requirements.txt
# requirements.txt
praw>=7.7.0
openai>=1.3.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
apscheduler>=3.10.1
requests>=2.31.0
pyyaml>=6.0
tabulate>=0.9.0
